fix: restrcture and typo
