+++ src/lib/services/wasm/code/index.ts
@@ -11,7 +11,16 @@ import type { Code, CodeData, CodesResponse } from "lib/services/types";
 import type { BechAddr, BechAddr20, Option } from "lib/types";
 
 import { getCodeData, getCodes, getCodesByAddress } from "./api";
-import { getCodeLcd, getCodesLcd } from "./lcd";
+import { getCodeLcd, getCodesLcd, getUploadAccessParamsLcd } from "./lcd";
+
+export const useUploadAccessParams = () => {
+  const endpoint = useLcdEndpoint();
+  return useQuery(
+    [CELATONE_QUERY_KEYS.UPLOAD_ACCESS_PARAMS, endpoint],
+    () => getUploadAccessParamsLcd(endpoint),
+    { keepPreviousData: true, refetchOnWindowFocus: false }
+  );
+};
 
 export const useCodes = (
   limit: number,
